true_project/requirements.txt
asgiref==3.7.2
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.2.1
djangorestframework==3.14.0
filelock==3.12.0
idna==3.4
Pillow==9.5.0
pipenv==2023.6.2
platformdirs==3.5.1
pycodestyle==2.10.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
virtualenv-clone==0.5.7
